Moved Mesh.java to subcomponents
